fix not in transaction exceptions
The transaction-related exceptions were due to retrieving the Model
outside of a READ or WRITE transaction. Now the namedModelURI, which is
null for the default model, is provided and the correct model is always
obtained within a transaction.

This meant that the union model must also be joined within a
transaction. The implementation suffered as the union boolean needed to
be passed into the RepositoryModel collaborator.