re-align RDF.rb abstractions with Jena
RDF::Jena::Repository now models a Jena DatasetGraph to act like a quad
store and container for graphs of triples (e.g. set of triples under a
graph).

RDF::Jena::Graph now models a Jena Graph. It is a set of triples within
either the default graph (:default) or a named graph.

implemented insert_reader taking an IO object, when Repository the IO is
read blindly as NQUADS, when Graph NTRIPLES.

implemented insert_statements taking an enumerator, when Repository the
enumerator is mapped to Jena Quad, when Graph it is mapped Jena Triple.

included more converters between RDF.rb and Jena.

version 0.3.0